Add vuu-auth package
Move the shared authentication contract and React provider into a dedicated UI package, and wire it into builds and publishing.
