+++ vuu-ui/packages/vuu-auth/README.md
@@ -1,0 +1,3 @@
+# @vuu-ui/vuu-auth
+
+Authentication contracts and React integration for VUU UI applications.
