+++ vuu-ui/packages/vuu-auth/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@vuu-ui/vuu-auth",
+  "version": "2.2.0",
+  "description": "VUU authentication APIs and React provider",
+  "main": "src/index.ts",
+  "author": "heswell",
+  "license": "Apache-2.0",
+  "type": "module",
+  "scripts": {
+    "build": "node ../../scripts/run-build-rslib.ts",
+    "type-defs": "node ../../scripts/build-type-defs.mjs"
+  },
+  "dependencies": {
+    "@vuu-ui/vuu-data-remote": "2.2.0",
+    "@vuu-ui/vuu-utils": "2.2.0"
+  },
+  "peerDependencies": {
+    "react": "^19.2.3"
+  },
+  "sideEffects": false
+}
