+++ vuu-ui/packages/vuu-data-remote/src/index.ts
@@ -10,7 +10,6 @@ export * from "./data-source";
 export { LostConnectionHandler, RetryOptions } from "./LostConnectionHandler";
 export * from "./message-utils";
 export { VuuAuthenticator, VuuAuthTokenIssuePolicy } from "./auth/VuuAuthenticator";
-export { VuuAuthProvider, type AuthProviderClass } from "./auth/VuuAuthProvider";
-export type { AuthProvider, User } from "./auth/AuthProvider";
+export { VuuAuthProvider } from "./auth/VuuAuthProvider";
 export * from "./VuuDataSource";
 export { isConnected, type ConnectionStatus } from "./WebSocketConnection";
