+++ vuu-ui/packages/vuu-shell/package.json
@@ -19,6 +19,7 @@
     "@salt-ds/core": "1.67.0",
     "@salt-ds/styles": "0.2.1",
     "@salt-ds/window": "0.1.1",
+    "@vuu-ui/vuu-auth": "2.2.0",
     "@vuu-ui/vuu-context-menu": "2.2.0",
     "@vuu-ui/vuu-data-react": "2.2.0",
     "@vuu-ui/vuu-data-remote": "2.2.0",
