+++ vuu-ui/packages/vuu-shell/src/shell.tsx
@@ -1,5 +1,6 @@
 import { useComponentCssInjection } from "@salt-ds/styles";
 import { useWindow } from "@salt-ds/window";
+import { useLoggedInUser } from "@vuu-ui/vuu-auth";
 import { ContextMenuProvider } from "@vuu-ui/vuu-context-menu";
 import type { LayoutChangeHandler } from "@vuu-ui/vuu-layout";
 import { LayoutProvider, StackLayout } from "@vuu-ui/vuu-layout";
@@ -30,7 +31,6 @@ import {
 } from "./workspace-management";
 import { loadingJSON } from "./workspace-management/defaultWorkspaceJSON";
 import { useLostConnection } from "@vuu-ui/vuu-data-react";
-import { useLoggedInUser } from "./authentication-provider/AuthenticationProvider";
 
 import shellCss from "./shell.css";
