+++ vuu-ui/sample-apps/app-vuu-example/package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@fontsource/open-sans": "^4.5.13",
     "@salt-ds/core": "1.67.0",
+    "@vuu-ui/vuu-auth": "2.2.0",
     "@vuu-ui/vuu-data-local": "2.2.0",
     "@vuu-ui/vuu-data-remote": "2.2.0",
     "@vuu-ui/vuu-table-types": "2.2.0",
