+++ vuu-ui/scripts/build-all-rslib.ts
@@ -31,6 +31,7 @@ export const buildAll = async () => {
     "vuu-ui-controls",
   ];
   const wave3 = [
+    "vuu-auth",
     "vuu-filters",
     "vuu-popups"
   ];
