+++ vuu-ui/scripts/publish.mjs
@@ -3,6 +3,7 @@ import { execWait, getCommandLineArg } from "./utils.mjs";
 const debug = getCommandLineArg("--debug");
 
 const packages = [
+  "vuu-auth",
   "vuu-chart",
   "vuu-codemirror",
   "vuu-context-menu",
